Renamed module, added enum
Renamed the module to 'atmotube' which is easier to remember. Also added an enum for the GATT characteristic UUIDs, so it is easier to use. Updated the examples and test cases to match.

A future to do is to come up with a test bleak client so I can test the functionality without having to be at my desk connecting to my particular atmotube.